requirements.txt
You need to install the following
tensorflow 1.0
python 3.5
pip 10
mongodb 4.0 (community)
tweepy 3.6
jsonpickle 0.9
tweet-preprocessor
preprocessing 0.1.13
pymongo
fix-yahoo-finance
pandas_datareader